+++ src/tests/handle_error.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::routing::BoxRoute;
 use std::future::{pending, ready};
 use tower::{timeout::TimeoutLayer, MakeService};
 
@@ -174,12 +175,12 @@ fn layered() {
     check_make_svc::<_, _, _, Infallible>(app.into_make_service());
 }
 
-#[tokio::test] // async because of `.boxed()`
-async fn layered_boxed() {
+#[test]
+fn layered_boxed() {
     let app = Router::new()
         .route("/echo", get::<_, Body, _>(unit))
         .layer(timeout())
-        .boxed()
+        .layer(BoxRoute::<Body, BoxError>::layer())
         .handle_error(handle_error::<BoxError>);
 
     check_make_svc::<_, _, _, Infallible>(app.into_make_service());